add deep_wasm_eval for deepvm frontend
