+++ src/vm/include/deep_wasm.h
@@ -1,0 +1,36 @@
+/*
+ * Copyright (C) 2019 Intel Corporation.  All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+ */
+/**
+ * @file   bh_log.h
+ * @date   Tue Nov  8 18:19:10 2011
+ *
+ * @brief This log system supports wrapping multiple outputs into one
+ * log message.  This is useful for outputting variable-length logs
+ * without additional memory overhead (the buffer for concatenating
+ * the message), e.g. exception stack trace, which cannot be printed
+ * by a single log calling without the help of an additional buffer.
+ * Avoiding additional memory buffer is useful for resource-constraint
+ * systems.  It can minimize the impact of log system on applications
+ * and logs can be printed even when no enough memory is available.
+ * Functions with prefix "_" are private functions.  Only macros that
+ * are not start with "_" are exposed and can be used.
+ */
+
+#ifndef _DEEP_WASM_H
+
+#include "bh_platform.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+int deep_wasm_vm_init (void);
+int deep_wasm_eval (uint8 * wasm_buf, uint32 wasm_size);
+
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif  /* _DEEP_WASM_H */
